Store sum in BigInt to avoid int overflow.

+++ src/main/scala/sl/challenge/Solver.scala
@@ -4,7 +4,7 @@ import sl.challenge.model.Result
 import sl.challenge.model.Types.{Path, Triangle}
 
 object Solver {
-  private type Ordering = (Int, Int) => Boolean
+  private type Ordering = (BigInt, BigInt) => Boolean
 
   private def MIN: Ordering = _ < _
