+++ src/main/scala/sl/challenge/model/Result.scala
@@ -1,3 +1,3 @@
 package sl.challenge.model
 
-case class Result(path: List[Int], val sum : Int)
+case class Result(path: List[Int], val sum : BigInt)
